+++ YMMProjectManager/Presentation/Timeline/YmmTimelineVisualSafetyGuard.cs
@@ -1,0 +1,7 @@
+namespace YMMProjectManager.Presentation.Timeline;
+
+public static class YmmTimelineVisualSafetyGuard
+{
+    public static bool IsVisualAttachAllowed(PureTimelineExperimentalOptions options)
+        => !options.ForbidVisualTreeAttach;
+}